+++ regionmask/defined_regions/__init__.py
@@ -1,9 +1,9 @@
 # flake8: noqa
 
-from . import _ressources
+from . import _natural_earth, _ressources
 from ._ar6 import ar6
 from ._ar6_pre_revisions import _ar6_pre_revisions
+from ._natural_earth import natural_earth, natural_earth_v4_1_0, natural_earth_v5_0_0
 from .giorgi import giorgi
-from .natural_earth import _maybe_get_column, _obtain_ne, natural_earth
 from .prudence import prudence
 from .srex import srex